requirements.txt
fastapi
uvicorn[standard]
python-docx
pdfplumber
mammoth
python-multipart
textract
